Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 int main() {
-    int a,b;
+    float a,b;
     char opt;
-    scanf("%d %d %c",&a,&b,&opt);
+    scanf("%f %f %c",&a,&b,&opt);
     switch(opt) {
         case '+':
-        printf("%d",a+b);
+        printf("%.0f",a+b);
         break;
         case '-':
-        printf("%d",a-b);
+        printf("%.0f",a-b);
         break;
         case '*':
-        printf("%d",a*b);
+        printf("%.0f",a*b);
         break;
         case '/':
-        printf("%d",a/b);
+        printf("%.0f",a/b);
         break;
         default:
         printf("error");
